+++ sc/social/like/upgrades/v3040/__init__.py
@@ -17,48 +17,76 @@ def update_configlet_information(setup_tool):
         logger.info('Configlet information updated')
 
 
-def _enforce_type_constraints(portal_types):
-    """Return ReallyUserFriendlyTypes only. Old field values were not
-    restricted in any way; new field values must be terms of this
-    vocabulary.
-    """
-    from zope.component import getUtility
-    from zope.schema.interfaces import IVocabularyFactory
-    name = 'plone.app.vocabularies.ReallyUserFriendlyTypes'
-    friendly_types = getUtility(IVocabularyFactory, name)(None)
-    return tuple([i for i in portal_types if i in friendly_types])
-
-
-def migrate_settings_to_registry(setup_tool):
+def migrate_settings_to_registry(setup_tool):  # noqa: C901
     """Migrate settings to registry."""
+
+    def filter_types(portal_types):
+        """Return ReallyUserFriendlyTypes only. Old field values were
+        not restricted in any way; new field values must be terms of
+        this vocabulary.
+        """
+        from zope.component import getUtility
+        from zope.schema.interfaces import IVocabularyFactory
+        name = 'plone.app.vocabularies.ReallyUserFriendlyTypes'
+        friendly_types = getUtility(IVocabularyFactory, name)(None)
+        return tuple([i for i in portal_types if i in friendly_types])
+
+    def safe_migrate(old_attr, new_attr=None, to_string=False):
+        """Copy value for property sheet to registry record avoiding
+        AttributeError; rename record if needed. In case of error
+        the record must already have loaded its default value.
+        """
+        try:
+            value = getattr(old_props, old_attr)
+        except AttributeError:
+            pass
+
+        if to_string:  # convert unicode to string?
+            # avoid 'None' on None values (missing value)
+            value = str(value) if value else ''
+
+        if new_attr is None:
+            new_attr = old_attr
+        setattr(settings, new_attr, value)
+
     profile = 'profile-{0}:default'.format(PROJECTNAME)
     setup_tool.runImportStepFromProfile(profile, 'plone.app.registry')
     portal_properties = api.portal.get_tool(name='portal_properties')
-    if 'sc_social_likes_properties' in portal_properties:
-        old_props = portal_properties.sc_social_likes_properties
-        registry = getUtility(IRegistry)
-        settings = registry.forInterface(ISocialLikeSettings)
-        # ignore types not allowed
-        types_ = _enforce_type_constraints(old_props.enabled_portal_types)
-        settings.enabled_portal_types = types_
-        settings.plugins_enabled = old_props.plugins_enabled
-        settings.typebutton = old_props.typebutton
-        settings.do_not_track = old_props.do_not_track
-        # this property may have an invalid value under certain circunstances
-        try:
-            settings.fbaction = old_props.fbaction
-        except ConstraintNotSatisfied:  # empty string
-            settings.fbaction = u'like'  # use default
-        settings.fbbuttons = old_props.fbbuttons
-        # these fields are no longer TextLine but ASCIILine
-        # we need to avoid 'None' values caused by missing defaults
-        settings.facebook_username = (
-            str(old_props.fbadmins) if old_props.fbadmins else '')
-        settings.facebook_app_id = (
-            str(old_props.fbapp_id) if old_props.fbapp_id else '')
-        settings.twitter_username = (
-            str(old_props.twittvia) if old_props.twittvia else '')
-        logger.info('Settings migrated')
+    if 'sc_social_likes_properties' not in portal_properties:
+        logger.warn('Property sheet not found; using defaults')
+        return
 
-        del portal_properties['sc_social_likes_properties']
-        logger.info('Property sheet "sc_social_likes_properties" removed')
+    old_props = portal_properties.sc_social_likes_properties
+    registry = getUtility(IRegistry)
+    settings = registry.forInterface(ISocialLikeSettings)
+
+    # ignore types not allowed
+    try:
+        portal_types = old_props.enabled_portal_types
+    except AttributeError:
+        pass
+    else:
+        portal_types = filter_types(portal_types)
+        settings.enabled_portal_types = portal_types
+
+    safe_migrate('plugins_enabled')
+    safe_migrate('typebutton')
+    safe_migrate('do_not_track')
+
+    try:
+        safe_migrate('fbaction')
+    except ConstraintNotSatisfied:
+        pass  # skip on empty string
+
+    safe_migrate('fbbuttons')
+
+    # these fields are no longer TextLine but ASCIILine
+    # and have a different name
+    safe_migrate('fbadmins', 'facebook_username', to_string=True)
+    safe_migrate('fbapp_id', 'facebook_app_id', to_string=True)
+    safe_migrate('twittvia', 'twitter_username', to_string=True)
+
+    logger.info('Settings migrated')
+
+    del portal_properties['sc_social_likes_properties']
+    logger.info('Property sheet removed')